# docs/api-contract.openapi.yaml
openapi: 3.0.3
info:
  title: Jiwambe Onboarding — Platform API
  description: |
    Upstream API consumed by **jiwambe-onboarding-app** via its Next.js BFF.

    - Human-readable contract: [`api-contract.md`](./api-contract.md)
    - BFF routes mirror these paths under `/api/*` (see markdown BFF map)
    - Money fields are server-computed; clients display only
  version: 1.0.0
  contact:
    name: Jiwambe Platform
servers:
  - url: "{jiwambeApiBaseUrl}"
    description: Includes `/api/v1` — set via `JIWAMBE_API_BASE_URL`
    variables:
      jiwambeApiBaseUrl:
        default: https://api.example.jiwambe.com/api/v1
...
security:
  - bearerAuth: []
paths:
...
  /onboarding/auth/otp/resend:
    post:
      tags: [OfficerAuth]
      summary: Resend SMS OTP
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [otp_session_id]
              properties:
                otp_session_id:
                  type: string
      responses:
        "200":
          description: Code resent
          content:
            application/json:
              schema:
                type: object
                properties:
                  resend_available_in_seconds:
                    type: integer
                  message:
                    type: string
        "410":
          $ref: "#/components/responses/OtpExpired"
        "429":
          $ref: "#/components/responses/RateLimited"
...
  /onboarding/auth/activate:
    post:
      tags: [OfficerAuth]
      summary: Validate CRM activation magic-link token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
      responses:
        "200":
          description: Activation session started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivationSessionResponse"
        "410":
          $ref: "#/components/responses/ActivationExpired"
  /onboarding/auth/activate/password:
    post:
      tags: [OfficerAuth]
      summary: Set password after activation
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [activation_session_id, password]
              properties:
                activation_session_id:
                  type: string
                password:
                  type: string
                  minLength: 8
      responses:
        "200":
          description: Password set
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
        "410":
          $ref: "#/components/responses/ActivationExpired"
...
  /onboarding/auth/password/change:
    post:
      tags: [OfficerAuth]
      summary: Change password while signed in
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, password, password_confirm]
              properties:
                current_password:
                  type: string
                password:
                  type: string
                  minLength: 8
                password_confirm:
                  type: string
                  minLength: 8
      responses:
        "200":
          description: Password updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
        "400":
          description: weak_password or invalid_body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "401":
          description: invalid_credentials (wrong current password) or missing bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
...
  /onboarding/applications/{id}:
    parameters:
      - $ref: "#/components/parameters/ApplicationId"
    get:
      tags: [Applications]
      summary: Get full application
      responses:
        "200":
          description: Application resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationEnvelope"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Applications]
      summary: Partial update (requires version)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationPatch"
      responses:
        "200":
          description: Updated application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationEnvelope"
        "409":
          $ref: "#/components/responses/VersionConflict"
        "404":
          $ref: "#/components/responses/NotFound"
  /onboarding/applications/{id}/pause:
    parameters:
      - $ref: "#/components/parameters/ApplicationId"
    post:
      tags: [Applications]
      summary: Pause capture
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReasonRequest"
      responses:
        "200":
          description: Paused
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationEnvelope"
...
  /onboarding/applications/{id}/disqualify:
    parameters:
      - $ref: "#/components/parameters/ApplicationId"
    post:
      tags: [Applications]
      summary: Disqualify (DRAFT or PAUSED only)
      description: See onboarding-disqualify-rules.md
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReasonRequest"
      responses:
        "200":
          description: Disqualified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationEnvelope"
  /onboarding/applications/{id}/documents/init:
    parameters:
      - $ref: "#/components/parameters/ApplicationId"
    post:
      tags: [Applications]
      summary: Init presigned document upload
      x-implemented: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [purpose, contentType, byteSize]
              properties:
                purpose:
                  type: string
                  example: id_front
                contentType:
                  type: string
                byteSize:
                  type: integer
      responses:
        "200":
          description: Upload instructions
          content:
            application/json:
              schema:
                type: object
                properties:
                  documentId:
                    type: string
                  uploadUrl:
                    type: string
                    format: uri
                  uploadHeaders:
                    type: object
                    additionalProperties:
                      type: string
                  expiresAt:
                    type: string
                    format: date-time
...
  /customers/search:
    post:
      tags: [Customers]
      summary: Search portal leads by phone or national ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerSearchRequest"
      responses:
        "200":
          description: Matches
          content:
            application/json:
              schema:
                type: object
                required: [matches]
                properties:
                  matches:
                    type: array
                    items:
                      $ref: "#/components/schemas/CustomerSearchMatch"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ApplicationId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Application reference code or internal id
  responses:
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
...
    RateLimited:
      description: Rate limited
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ErrorBody"
              - type: object
                properties:
                  retry_after_seconds:
                    type: integer
    OtpExpired:
      description: OTP session expired
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    ActivationExpired:
      description: Activation token expired
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
...
    VersionConflict:
      description: Optimistic lock conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
...
  schemas:
...
    ErrorBody:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
...
    ActivationSessionResponse:
      type: object
      properties:
        activation_session_id:
          type: string
        email_masked:
          type: string
...
    LifecycleState:
      type: string
      enum:
        - DRAFT
        - PAUSED
        - OPS_REVIEW
        - LMS_CREATED
        - AGREEMENT_SIGNED
        - READY_FOR_RELEASE
        - ACTIVE_LOAN
        - DISQUALIFIED
...
    ReasonRequest:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
          minLength: 6
...
    ApplicationResource:
      type: object
      description: Full resource — see sample-application-resource.ts
      properties:
        referenceCode:
          type: string
        version:
          type: integer
        lifecycleState:
          $ref: "#/components/schemas/LifecycleState"
        customer:
          type: object
        financing:
          type: object
        operatingModel:
          type: object
        bikeAssignment:
          type: object
          nullable: true
    ApplicationPatch:
      type: object
      required: [version]
      properties:
        version:
          type: integer
        customer:
          type: object
        financing:
          type: object
        operatingModel:
          type: object
        bikeAssignment:
          type: object
    ApplicationEnvelope:
      type: object
      required: [application]
      properties:
        application:
          $ref: "#/components/schemas/ApplicationResource"
    CustomerSearchRequest:
      type: object
      properties:
        phone:
          type: string
        nationalId:
          type: string
          nullable: true
    CustomerSearchMatch:
      type: object
      properties:
        leadId:
          type: string
        source:
          type: string
        displayName:
          type: string
        phoneMasked:
          type: string
        nationalIdMasked:
          type: string
...